update: address PR comments
